rename insights to marketPriceInsights move loadBatchPriceInsights update tests
