+++ src/lib/loadBatchPriceInsights.ts
